moved all major JS out to init.js file, which is loaded separately
